refactor(controllers): reorganize validation and response utilities
Also remove unused code such as appError.js and helper functions
