--- backend/src/utils/appError.js
@@ -1,8 +1,0 @@
-export const AppError = class extends Error {
-  constructor(message, statusCode) {
-    super(message);
-    this.statusCode = statusCode;
-    this.isOperational = true;
-    Error.captureStackTrace(this, this.constructor);
-  }
-};
