--- backend/src/utils/helpers.js
@@ -1,4 +1,0 @@
-export const isValidEmail = (email) => {
-  const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
-  return emailRegex.test(email);
-};
